codeArtifact rename to code_artifact
